make HydraConsolePainter an inner class
